% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glances.R
\name{glance.melded}
\alias{glance.melded}
\title{Summarise Amelia's multiple imputations}
\usage{
\method{glance}{melded}(x, ...)
}
\arguments{
\item{x}{Amelia melded imputed data}

\item{...}{other arguments passed to methods}
}
\value{
one line dataframe summary of the multiply imputed data
}
\description{
A broom glance summary for multiple imputations from Andrew's blog post:
https://www.andrewheiss.com/blog/2018/03/07/amelia-tidy-melding/
Andrews note: This means properly melded parameters and the simple average
of the parameters of these models are roughly the same sake of simplicty
we just take the average here.
}
