% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs.melded}
\alias{nobs.melded}
\title{Take the number of observations from the first Amelia melded model}
\usage{
\method{nobs}{melded}(object, ...)
}
\arguments{
\item{object}{Amelia melded model}

\item{...}{other options to pass to `nobs`}
}
\value{
integer
}
\description{
Take the number of observations from the first Amelia melded model
}
