% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.melded}
\alias{tidy.melded}
\title{Tidy summaries for melded data}
\usage{
\method{tidy}{melded}(x, ..., conf.int = FALSE, conf.level = 0.95)
}
\arguments{
\item{x}{melded object from Amelia}

\item{...}{other arguments passed to methods}

\item{conf.int}{Return confidence intervals. Default is FALSE}

\item{conf.level}{Confidence level (if confidence intervals are use).
Default is 0.95.}
}
\value{
tidy dataframe with pooled estimates from imputed data
}
\description{
This is an approach for tidying Amelia imputed dataframes from Andrew Heiss
  <https://www.andrewheiss.com/blog/2018/03/08/amelia-broom-huxtable/>
}
